move tx checks to their own file, split up check_balance_for_total_fees
